Relocating app code
